+++ tests/run-clibrary/main.f
@@ -123,7 +123,7 @@ contains
 !    call function10("bar", 2.0d0)
 !    call assert_true(.true.)
 
-    call sum(5, [1,2,3,4,5], rv_int)
+    call sum([1,2,3,4,5], rv_int)
     call assert_true(rv_int .eq. 15)
 
 !    rv_int = typefunc(2)