+++ CHANGELOG.md
@@ -6,11 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support function pointer arguments.
+- Improve support for Python wrappers, including NumPy support.
+- Add implied attribute used to compute value of an argument.
+
 ### Fixed
 - Add continuations to long lines.
   This helps Fortran which has a line length limit.
-- Support function pointer arguments.
-- Improve support for Python wrappers, including NumPy support.
 
 ## v0.7.0 - 2018-01-22
 ### Added